+++ lua/obsidian/yaml/parser.lua
@@ -220,10 +220,10 @@ Parser._try_parse_field = function(self, lines, i, text)
   local _, key, value
 
   -- First look for start of mapping, array, block, etc, e.g. 'foo:'
-  _, _, key = string.find(text, "([a-zA-Z0-9_-]+):$")
+  _, _, key = string.find(text, "([a-zA-Z0-9_-]+[a-zA-Z0-9_ -]*):$")
   if not key then
     -- Then try inline field, e.g. 'foo: bar'
-    _, _, key, value = string.find(text, "([a-zA-Z0-9_-]+): (.*)")
+    _, _, key, value = string.find(text, "([a-zA-Z0-9_-]+[a-zA-Z0-9_ -]*): (.*)")
   end
 
   value = value and util.strip_whitespace(value) or nil